Refactor replication_gui functions to static methods
