+++ libraries/replication.inc.php
@@ -100,7 +100,7 @@ $slave_variables  = array(
  * define important variables, which need to be watched for
  * correct running of replication in slave mode
  *
- * @usedby PMA_getHtmlForReplicationStatusTable()
+ * @usedby PhpMyAdmin\ReplicationGui::getHtmlForReplicationStatusTable()
  */
 // TODO change to regexp or something, to allow for negative match.
 // To e.g. highlight 'Last_Error'
