+++ server_replication.php
@@ -5,6 +5,8 @@
  *
  * @package PhpMyAdmin
  */
+
+use PhpMyAdmin\ReplicationGui;
 use PhpMyAdmin\Response;
 use PhpMyAdmin\Server\Common;
 
@@ -13,9 +15,7 @@ use PhpMyAdmin\Server\Common;
  */
 require_once 'libraries/common.inc.php';
 require_once 'libraries/server_common.inc.php';
-
 require_once 'libraries/replication.inc.php';
-require_once 'libraries/replication_gui.lib.php';
 
 /**
  * Does the common work
@@ -46,7 +46,7 @@ if (isset($_REQUEST['url_params']) && is_array($_REQUEST['url_params'])) {
 /**
  * Handling control requests
  */
-PMA_handleControlRequest();
+ReplicationGui::handleControlRequest();
 
 /**
  * start output
@@ -55,19 +55,19 @@ $response->addHTML('<div id="replication">');
 $response->addHTML(Common::getHtmlForSubPageHeader('replication'));
 
 // Display error messages
-$response->addHTML(PMA_getHtmlForErrorMessage());
+$response->addHTML(ReplicationGui::getHtmlForErrorMessage());
 
 if ($GLOBALS['replication_info']['master']['status']) {
-    $response->addHTML(PMA_getHtmlForMasterReplication());
+    $response->addHTML(ReplicationGui::getHtmlForMasterReplication());
 } elseif (! isset($_REQUEST['mr_configure'])
     && ! isset($_REQUEST['repl_clear_scr'])
 ) {
-    $response->addHTML(PMA_getHtmlForNotServerReplication());
+    $response->addHTML(ReplicationGui::getHtmlForNotServerReplication());
 }
 
 if (isset($_REQUEST['mr_configure'])) {
     // Render the 'Master configuration' section
-    $response->addHTML(PMA_getHtmlForMasterConfiguration());
+    $response->addHTML(ReplicationGui::getHtmlForMasterConfiguration());
     exit;
 }
 
@@ -76,12 +76,12 @@ $response->addHTML('</div>');
 if (! isset($_REQUEST['repl_clear_scr'])) {
     // Render the 'Slave configuration' section
     $response->addHTML(
-        PMA_getHtmlForSlaveConfiguration(
+        ReplicationGui::getHtmlForSlaveConfiguration(
             $GLOBALS['replication_info']['slave']['status'],
             $server_slave_replication
         )
     );
 }
 if (isset($_REQUEST['sl_configure'])) {
-    $response->addHTML(PMA_getHtmlForReplicationChangeMaster("slave_changemaster"));
+    $response->addHTML(ReplicationGui::getHtmlForReplicationChangeMaster("slave_changemaster"));
 }
