+++ server_status.php
@@ -18,7 +18,6 @@ require_once 'libraries/server_common.inc.php';
  * Replication library
  */
 require_once 'libraries/replication.inc.php';
-require_once 'libraries/replication_gui.lib.php';
 
 /**
  * start output
