+++ server_status_monitor.php
@@ -13,7 +13,6 @@ use PhpMyAdmin\Response;
 require_once 'libraries/common.inc.php';
 require_once 'libraries/server_common.inc.php';
 require_once 'libraries/replication.inc.php';
-require_once 'libraries/replication_gui.lib.php';
 
 $response = Response::getInstance();
