+++ server_status_variables.php
@@ -14,7 +14,6 @@ use PhpMyAdmin\Server\Status\Variables;
 require_once 'libraries/common.inc.php';
 require_once 'libraries/server_common.inc.php';
 require_once 'libraries/replication.inc.php';
-require_once 'libraries/replication_gui.lib.php';
 
 /**
  * flush status variables if requested
